+++ test/react-web/client/graphql/fragments.test.tsx
@@ -5,7 +5,7 @@ import ApolloClient from 'apollo-client';
 import { InMemoryCache as Cache } from 'apollo-cache-inmemory';
 import { mockSingleLink } from '../../../../src/test-utils';
 import { ApolloProvider, graphql } from '../../../../src';
-import '../../../setup/toEqualWithoutSymbol';
+import '../../../test-utils/toEqualWithoutSymbol';
 
 describe('fragments', () => {
   // XXX in a later version, we should support this for composition
