+++ test/test-utils/toEqualWithoutSymbol.ts
